backend/requirements.txt
Flask==2.0.1
requests==2.25.1
PyGithub==1.55
openai==0.27.0
